+++ vertx-example/src/main/kotlin/uy/kohesive/kovert/vertx/sample/web/PublicWebController.kt
@@ -1,0 +1,16 @@
+package uy.kohesive.kovert.vertx.sample.web
+
+import uy.kohesive.kovert.core.HttpRedirect
+import uy.kohesive.kovert.core.Rendered
+import uy.kohesive.kovert.vertx.sample.services.User
+
+public class PublicWebController {
+    @Rendered("login.html.ftl")
+    val viewLogin = fun PublicAccess.(): EmptyModel {
+        if (user != null) {
+            throw HttpRedirect("/app") // TODO: replace with looking up HREF from a controller
+        }
+        return EmptyModel()
+    }
+}
+