+++ vertx-jdk8/src/test/kotlin/uy/kohesive/kovert/vertx/boot/TestKovertApp.kt
@@ -40,21 +40,21 @@ class TestKovertApp {
             val client = deployment.vertx.createHttpClient(HttpClientOptions().setDefaultHost("localhost").setDefaultPort(Injekt.get<KovertVerticleConfig>().listeners.first().port))
 
             val frankJson = """{"id":1,"name":"Frank","age":30}"""
-            client.testServer(HttpMethod.GET, "person/1", assertResponse = frankJson)
-            client.testServer(HttpMethod.GET, "person/id/1", assertResponse = frankJson)
-            client.testServer(HttpMethod.GET, "person/name/frank", assertResponse = """[$frankJson]""")
+            client.testServer(HttpMethod.GET, "api/person/1", assertResponse = frankJson)
+            client.testServer(HttpMethod.GET, "api/person/id/1", assertResponse = frankJson)
+            client.testServer(HttpMethod.GET, "api/person/name/frank", assertResponse = """[$frankJson]""")
 
-            client.testServer(HttpMethod.GET, "person/id/1991991", assertStatus = 404)
-            client.testServer(HttpMethod.GET, "person/name/doogie", assertStatus = 404)
+            client.testServer(HttpMethod.GET, "api/person/id/1991991", assertStatus = 404)
+            client.testServer(HttpMethod.GET, "api/person/name/doogie", assertStatus = 404)
 
             client.testServer(HttpMethod.GET, "/index.html", assertResponse = """<html><head></head><body>Shhhh!</body></html>""")
 
             client.testServer(HttpMethod.GET, "/something/funky", assertResponse = "\"Funky\"")
 
             val collokiaJson = """{"name":"Collokia","country":"Uruguay"}"""
-            client.testServer(HttpMethod.GET, "company/Collokia", assertResponse = """{"status":"OK","data":$collokiaJson}""")
-            client.testServer(HttpMethod.GET, "company/name/Collokia", assertResponse = """{"status":"OK","data":$collokiaJson}""")
-            client.testServer(HttpMethod.GET, "company/country/Uruguay", assertResponse = """{"status":"OK","data":[$collokiaJson,{"name":"Bremeld","country":"Uruguay"}]}""")
+            client.testServer(HttpMethod.GET, "api/company/Collokia", assertResponse = """{"status":"OK","data":$collokiaJson}""")
+            client.testServer(HttpMethod.GET, "api/company/name/Collokia", assertResponse = """{"status":"OK","data":$collokiaJson}""")
+            client.testServer(HttpMethod.GET, "api/company/country/Uruguay", assertResponse = """{"status":"OK","data":[$collokiaJson,{"name":"Bremeld","country":"Uruguay"}]}""")
 
         } finally {
             deployment.vertx.promiseUndeploy(deployment.deploymentId).get()
@@ -95,8 +95,8 @@ public class KovertApp(val configFile: Path) {
         KovertConfig.addVerbAlias("find", HttpVerb.GET)
 
         val initControllers = fun Router.(): Unit {
-            bindController(PeopleController(), "/person")
-            bindController(CompanyController(), "/company")
+            bindController(PeopleController(), "/api/person")
+            bindController(CompanyController(), "/api/company")
             bindController(RootController(), "/")
         }